feat: 커스텀 훅으로 기능 이동
